Change to int

+++ project2/include/expression.h
@@ -8,7 +8,7 @@
 class Expression
 {
 public:
-    virtual double evaluate() = 0;
+    virtual int evaluate() = 0;
 };
 
 #endif //PROJECT2_EXPRESSION_H
